+++ tests/test_public_types.py
@@ -4,7 +4,7 @@ from urllib.parse import parse_qs, urlparse
 
 import responses
 
-from pyvworld import (
+from vworld import (
     AddressCategory,
     AddressType,
     BBox,